module #5. Locking Down DeleteItem Permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -41,7 +41,12 @@ const Mutations = {
   },
   deleteItem: async (_, args, ctx, info) => {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(p => ['ADMIN', 'ITEMDELETE'].includes(p));
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to do that!');
+    }
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   signup: async (_, args, ctx, info) => {
